// Voucher.UI/src/app/shared/model/accommodation-voucher.model.ts
export class AccommodationVoucher {
    id?: number;
    createdBy: string;
    canceledBy: string;
    createdDate: Date;
    validUntil: Date;
    canceledDate: Date;
    printedDate: Date;
    serviceProviderId: number;
    freeText: string;
    flightId: number;
    isActive: boolean;
    pseudoCityCode: string;
    idPassenger: string;
    passengerName: string;
    status: number;
    roomType: number;
    dailyAmount: number;
    reason: string;
}